+++ src/main/java/com/maxim/pojo/order/StockOrder.java
@@ -8,7 +8,7 @@ public class StockOrder extends Order{
     public Double dynamic_loss;
     public boolean partialOrder; // 当前订单是否为子单
 
-    public StockOrder(String symbol, Double vol, Double price, LocalDate create_date, LocalDateTime create_timestamp,
+    public StockOrder(String symbol, Integer vol, Double price, LocalDate create_date, LocalDateTime create_timestamp,
                       LocalDateTime min_order_timestamp, LocalDateTime max_order_timestamp, String reason){
         super(symbol, vol, price, create_date, create_timestamp, min_order_timestamp, max_order_timestamp, reason);
         this.partialOrder = false; // 默认当前订单为完整订单
