+++ src/main/java/com/maxim/pojo/position/FuturePosition.java
@@ -23,7 +23,7 @@ public class FuturePosition extends Position{
     public Double history_max;
     public Double history_min;
 
-    public FuturePosition(Double price, Double vol, Double margin_rate, LocalDateTime min_timestamp, LocalDateTime max_timestamp,
+    public FuturePosition(Double price, Integer vol, Double margin_rate, LocalDateTime min_timestamp, LocalDateTime max_timestamp,
                           Double static_profit, Double static_loss,
                           Double dynamic_profit, Double dynamic_loss) {
         super(price, vol, min_timestamp, max_timestamp);
